review: extract Header and Footer
