+++ web-registry/src/components/organisms/RegistryLayout/RegistryLayout.Footer.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
+import { URL_PRIVACY, URL_TERMS_SERVICE } from 'globals';
 
 import {
   Footer,
   FooterItemProps as FooterItem,
 } from 'web-components/lib/components/footer/footer-new';
 
-import { URL_PRIVACY, URL_TERMS_SERVICE } from '../../globals';
-import { Link, RegistryIconLink } from '../atoms';
+import { Link, RegistryIconLink } from 'components/atoms';
 
-const AppFooter: React.FC<React.PropsWithChildren<unknown>> = () => {
+const RegistryLayoutFooter: React.FC<React.PropsWithChildren<unknown>> = () => {
   const { pathname } = useLocation();
   const isHidden = ['/project-pages'].some(route => pathname.startsWith(route));
 
@@ -81,4 +81,4 @@ const AppFooter: React.FC<React.PropsWithChildren<unknown>> = () => {
   );
 };
 
-export { AppFooter };
+export { RegistryLayoutFooter };
